Per the LLVM spec, indicies into struct fields must be i32 integers.
This changes the scheme to use i32s for all GEP indicies converted from Python ints, unless the index is too large to represent with an i32, in which case a wider integer is used.

+++ llvmlite/tests/test_ir.py
@@ -847,7 +847,7 @@ my_block:
         self.check_block(block, """\
             my_block:
                 %"c" = alloca {i32, i32}*
-                %"d" = getelementptr {i32, i32}*, {i32, i32}** %"c", i32 5, i32 %".1", i8 1
+                %"d" = getelementptr {i32, i32}*, {i32, i32}** %"c", i32 5, i32 %".1", i32 1
             """)
         # XXX test with more complex types
